+++ src/model.js
@@ -15,10 +15,11 @@ const rewind = require('@mapbox/geojson-rewind');
 const ElasticConnectUtil = require('./utils/elasticConnectUtil');
 const {GeoTileAggregation} = require('./subLayers/geoTileAggregation');
 const {GeoHashAggregation} = require('./subLayers/geoHashAggregation');
+const {GeoHexAggregation} = require('./subLayers/geoHexAggregation');
 
 module.exports = function (koop) {
     this.customSymbolizers = [];
-    this.customSubLayers = [new GeoTileAggregation(), new GeoHashAggregation()];
+    this.customSubLayers = [new GeoTileAggregation(), new GeoHashAggregation(), new GeoHexAggregation()];
     this.customIndexNameBuilder = undefined;
 
     this.setTimeExtent = function (featureCollection) {
@@ -93,6 +94,10 @@ module.exports = function (koop) {
                 maxRecordCount: indexConfig.maxResults,
                 extent: extent,
                 supportsPagination: false
+            },
+            filtersApplied: {
+                where: true,
+                geometry: true
             }
         };
 
@@ -132,7 +137,6 @@ module.exports = function (koop) {
                         this.setTimeExtent(featureCollection);
                     }
                 } catch (e) {
-                    console.error(e);
                     callback(e, featureCollection);
                 }
             }
@@ -258,16 +262,7 @@ module.exports = function (koop) {
                 // This does not appear to fix the "requested provider has no "idField" assignment. This can cause errors in ArcGIS clients" error.
                 //featureCollection.metadata.idField = "OBJECTID";
 
-                if (!featureCollection.filtersApplied) {
-                    featureCollection.filtersApplied = {};
-                }
 
-                featureCollection.filtersApplied.where = true;
-
-                // we have already filtered the geometries using the incoming request. there are bugs regarding
-                // filtering by geometry, so we can turn it off and use elastic search's filtering.
-
-                featureCollection.filtersApplied.geometry = true;
                 // featureCollection.filtersApplied.projection = true;
 
                 if (indexConfig.geometryType === "geo_point" && indexConfig.allowMultiPoint === true) {
@@ -279,7 +274,9 @@ module.exports = function (koop) {
                 }
 
                 // returnCountOnly only set. This appears to trigger the count response in featureserver
-                featureCollection.count = searchResponse.hits.hits.length;
+                if(!featureCollection.count){
+                    featureCollection.count = searchResponse.hits.hits.length;
+                }
 
                 // if there an offset
                 if (offset > 0) {
@@ -326,7 +323,6 @@ module.exports = function (koop) {
                         // logger.debug(`Total Time: ${(Date.now().valueOf() - startMillis)/1000} seconds`);
                         callback(null, subLayerFeatureCollection);
                     }).catch(error => {
-                        logger.error(error);
                         callback(error, featureCollection);
                     });
                 }).catch(error => {