+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added 
+* Support for geohex_aggregation on Elastic Clusters with a hex aggregation license
+
 ## [3.3.5] - 04-21-2022
 ### Fixed
 * Set supportsPagination to false since it is not supported